+++ src/comments/dto/update-comment.dto.ts
@@ -1,0 +1,12 @@
+import { Transform } from 'class-transformer';
+import { IsNotEmpty, IsString, MaxLength } from 'class-validator';
+
+import { trim } from '../../common/transformers/trim.transformer';
+
+export class UpdateCommentDto {
+  @Transform(trim)
+  @IsString()
+  @IsNotEmpty()
+  @MaxLength(2000)
+  content: string;
+}