+++ src/comments/types/comment-response.type.ts
@@ -1,0 +1,10 @@
+export type CommentResponse = {
+  id: string;
+  content: string;
+  createdAt: Date;
+  editedAt: Date | null;
+  user: { id: string; username: string };
+  likeCount: number;
+  replyCount: number;
+  liked: boolean;
+};